+++ hutool-core/src/main/java/cn/hutool/core/util/CharUtil.java
@@ -166,8 +166,24 @@ public class CharUtil {
 	 * @param ch 被检查的字符
 	 * @return true表示为数字字符，数字字符指0~9
 	 */
-	public static boolean isNumber(final char ch) {
+	public static boolean isNumber(char ch) {
 		return ch >= '0' && ch <= '9';
+	}
+	
+	/**
+	 * 是否为16进制规范的字符，判断是否为如下字符
+	 * <pre>
+	 * 1. 0~9
+	 * 2. a~f
+	 * 4. A~F
+	 * </pre>
+	 * 
+	 * @param c 字符
+	 * @return 是否为16进制规范的字符
+	 * @since 4.1.5
+	 */
+	public static boolean isHexChar(char c) {
+		return isNumber(c) || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F');
 	}
 
 	/**