+++ hutool-core/src/main/java/cn/hutool/core/util/EscapeUtil.java
@@ -1,5 +1,8 @@
 package cn.hutool.core.util;
 
+import cn.hutool.core.text.escape.Html4Escape;
+import cn.hutool.core.text.escape.Html4Unescape;
+
 /**
  * 转义和反转义工具类Escape / Unescape<br>
  * escape采用ISO Latin字符集对指定的字符串进行编码。<br>
@@ -8,19 +11,43 @@ package cn.hutool.core.util;
  * @author xiaoleilu
  */
 public class EscapeUtil {
-	
+
+	/**
+	 * 转义HTML4中的特殊字符
+	 * 
+	 * @param html HTML文本
+	 * @return 转义后的文本
+	 * @since 4.1.5
+	 */
+	public static String escapeHtml4(String html) {
+		Html4Escape escape = new Html4Escape();
+		return escape.replace(html).toString();
+	}
+
+	/**
+	 * 反转义HTML4中的特殊字符
+	 * 
+	 * @param html HTML文本
+	 * @return 转义后的文本
+	 * @since 4.1.5
+	 */
+	public static String unescapeHtml4(String html) {
+		Html4Unescape unescape = new Html4Unescape();
+		return unescape.replace(html).toString();
+	}
+
 	/**
 	 * Escape编码（Unicode）<br>
 	 * 该方法不会对 ASCII 字母和数字进行编码，也不会对下面这些 ASCII 标点符号进行编码： * @ - _ + . / 。其他所有的字符都会被转义序列替换。
-
+	 * 
 	 * @param content 被转义的内容
 	 * @return 编码后的字符串
 	 */
 	public static String escape(String content) {
-		if(StrUtil.isBlank(content)) {
+		if (StrUtil.isBlank(content)) {
 			return content;
 		}
-		
+
 		int i;
 		char j;
 		StringBuilder tmp = new StringBuilder();
@@ -32,7 +59,7 @@ public class EscapeUtil {
 
 			if (Character.isDigit(j) || Character.isLowerCase(j) || Character.isUpperCase(j)) {
 				tmp.append(j);
-			}else if (j < 256) {
+			} else if (j < 256) {
 				tmp.append("%");
 				if (j < 16) {
 					tmp.append("0");
@@ -48,14 +75,15 @@ public class EscapeUtil {
 
 	/**
 	 * Escape解码
+	 * 
 	 * @param content 被转义的内容
 	 * @return 解码后的字符串
 	 */
 	public static String unescape(String content) {
-		if(StrUtil.isBlank(content)) {
+		if (StrUtil.isBlank(content)) {
 			return content;
 		}
-		
+
 		StringBuilder tmp = new StringBuilder(content.length());
 		int lastPos = 0, pos = 0;
 		char ch;
@@ -83,9 +111,10 @@ public class EscapeUtil {
 		}
 		return tmp.toString();
 	}
-	
+
 	/**
 	 * 安全的unescape文本，当文本不是被escape的时候，返回原文。
+	 * 
 	 * @param content 内容
 	 * @return 解码后的字符串，如果解码失败返回原字符串
 	 */
@@ -93,7 +122,7 @@ public class EscapeUtil {
 		try {
 			return unescape(content);
 		} catch (Exception e) {
-			//Ignore Exception
+			// Ignore Exception
 		}
 		return content;
 	}